MS-631 IVFSQ8H Index support
Former-commit-id: 21e17a20794e4fde31e79c4bbd4e26d46c79d886

+++ cpp/src/cache/CpuCacheMgr.cpp
@@ -32,7 +32,7 @@ CpuCacheMgr::CpuCacheMgr() {
     server::Config& config = server::Config::GetInstance();
     Status s;
 
-    int32_t cpu_cache_cap;
+    int64_t cpu_cache_cap;
     s = config.GetCacheConfigCpuCacheCapacity(cpu_cache_cap);
     if (!s.ok()) {
         SERVER_LOG_ERROR << s.message();
